Fix and new feature
I have implemented the simple NES controller, it works with the keyboard UP, DOWN, LEFT, RIGHT, W, X, A and B keys.
Using the nestest.nes rom, I finally fixed all the cpu problems and now the cpu is running perfectly.
Because I don't want to use the PixelGameEngine in the finished version, I started to create a second cmake profile which will use something else.
The Copyright was not absolutely right, so I modified it.