Sound Engine and APU implemantation
I have almost completely redone the SoundEngine because I was regenerating all buffer when calling a generate function. Now, it will only update the buffers.
The APU is still in WIP. It will now output a very horrible sound. I think it is trying to play the same note each time we call it. Some more work is going to be done on the Sound Engine. Maybe a dirty flag that allow regeneration only when a value has been modified ?